move reader reblog selection to its own component
